requirements.txt
Django==1.10.5
psycopg2==2.6.2
djangorestframework==3.5.4
Markdown==2.6.8
django_filter==1.0.1